requirements.txt
Django==5.2
python-decouple==3.8
Requests==2.32.3
gunicorn>=20.1.0
whitenoise==6.9.0
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2024.2